refactor: rename theme-dark to theme-resolver
