+++ src/common/lib/theme/index.ts
@@ -1,3 +1,3 @@
 export * from './provider';
-export * from './theme-dark';
+export * from './theme-resolver';
 export * from './use-theme';
